api-bot/requirements.txt
fastapi
python-telegram-bot
uvicorn 